backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.19
pyyaml==6.0.2
google-generativeai==0.8.4
pypinyin==0.53.0
pdfplumber==0.11.4
PyMuPDF==1.25.3
reportlab==4.2.5
cos-python-sdk-v5==1.9.32
apscheduler==3.10.4
aiofiles==24.1.0
